requirements.txt
feedparser
google-generativeai